+++ src/components/Jobs/components/index.js
@@ -1,3 +1,2 @@
 export * from './Job';
-export * from './PreviewsJobs';
 export * from './RunningJobs';