+++ test/bigtest/mocks/job-executions.js
@@ -1,110 +1,6 @@
-import { JOB_STATUSES } from '../../../src/utils/constants';
+import { JOB_STATUSES } from '../../../src/utils';
 
-const {
-  PREPARING_FOR_PREVIEW,
-  READY_FOR_PREVIEW,
-  RUNNING,
-} = JOB_STATUSES;
-
-const previewJobs = [
-  {
-    id: '469eba83-41d1-4161-bd1a-0f46d8764c6a',
-    hrId: 199482989,
-    subordinationType: 'PARENT_SINGLE',
-    jobProfileInfo: { name: 'Main bib jobs (MARC)' },
-    fileName: 'import_1.mrc',
-    sourcePath: 'import_1.mrc',
-    runBy: {
-      firstName: 'Marie',
-      lastName: 'Curie',
-    },
-    progress: {
-      current: 23,
-      total: 33,
-    },
-    startedDate: '2018-11-20T14:50:44.000',
-    uiStatus: PREPARING_FOR_PREVIEW,
-    status: 'PROCESSING_FINISHED',
-  },
-  {
-    id: '469eba83-41d1-4161-bd1a-0f46d876477t',
-    hrId: 182982110,
-    subordinationType: 'PARENT_SINGLE',
-    jobProfileInfo: { name: 'Library indexing' },
-    fileName: 'import_1.mrc',
-    sourcePath: 'import_1.mrc',
-    runBy: {
-      firstName: 'Marie',
-      lastName: 'Doe',
-    },
-    progress: {
-      current: 43,
-      total: 70,
-    },
-    startedDate: '2018-11-20T13:42:44.000',
-    uiStatus: PREPARING_FOR_PREVIEW,
-    status: 'PROCESSING_FINISHED',
-  },
-  {
-    id: '469eba83-41d1-4161-bd1a-0f46d342177e',
-    hrId: 182982111,
-    subordinationType: 'PARENT_SINGLE',
-    jobProfileInfo: { name: 'Indexing' },
-    fileName: 'import_1.mrc',
-    sourcePath: 'import_1.mrc',
-    runBy: {
-      firstName: 'Marie',
-      lastName: 'Edwards',
-    },
-    progress: {
-      current: 13,
-      total: 33,
-    },
-    startedDate: '2018-11-20T13:11:49.000',
-    uiStatus: PREPARING_FOR_PREVIEW,
-    status: 'PROCESSING_FINISHED',
-  },
-  {
-    id: '469eba83-41d1-4161-bd1a-0f46d866177d',
-    hrId: 182982930,
-    subordinationType: 'PARENT_SINGLE',
-    jobProfileInfo: { name: 'Library indexing' },
-    fileName: 'import_1.mrc',
-    sourcePath: 'import_1.mrc',
-    runBy: {
-      firstName: 'Oliver',
-      lastName: 'Clarke',
-    },
-    progress: {
-      current: 33,
-      total: 33,
-    },
-    startedDate: '2018-11-20T13:11:49.000',
-    completedDate: '2018-11-20T22:31:34.000',
-    uiStatus: READY_FOR_PREVIEW,
-    status: 'PROCESSING_FINISHED',
-  },
-  {
-    id: '469eba83-41d1-4161-bd1a-0f46d86999pd',
-    hrId: 182982220,
-    subordinationType: 'PARENT_SINGLE',
-    jobProfileInfo: { name: 'BIB Import from Boston' },
-    fileName: 'import_1.mrc',
-    sourcePath: 'import_1.mrc',
-    runBy: {
-      firstName: 'Taylor',
-      lastName: 'Clarke',
-    },
-    progress: {
-      current: 5000,
-      total: 5000,
-    },
-    startedDate: '2018-11-20T14:13:49.000',
-    completedDate: '2018-11-20T20:51:34.000',
-    uiStatus: READY_FOR_PREVIEW,
-    status: 'PROCESSING_FINISHED',
-  },
-];
+const { RUNNING } = JOB_STATUSES;
 
 const runningJobs = [
   {
@@ -166,8 +62,6 @@ const runningJobs = [
   },
 ];
 
-export const jobExecutions = [...previewJobs, ...runningJobs];
-
-export const PREVIEW_JOBS_LENGTH = previewJobs.length;
+export const jobExecutions = [...runningJobs];
 
 export const RUNNING_JOBS_LENGTH = runningJobs.length;