+++ src/components/MatchCriterion/view/ExistingRecordSections/index.js
@@ -1,1 +1,1 @@
-export * from './ExistingSectionFolio';
+export { default as ExistingSectionFolio } from './ExistingSectionFolio';