+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Update the Action profile UI: Create/Edit screen (UIDATIMP-1215)
 * Change UI flow for import jobs cancelled by users  (UIDATIMP-1173)
 * Update the Match profile UI: Create/Edit screen (UIDATIMP-1214)
+* Get rid of jobExecutions request for the Preview section (UIDATIMP-1239)
 
 ### Bugs fixed:
 * Invoice Field Mapping profile: Acq unit and Batch group dropdown lists should be in alphabetical order (UIDATIMP-1236)