+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Update the Match profile UI: Create/Edit screen (UIDATIMP-1214)
 * Get rid of jobExecutions request for the Preview section (UIDATIMP-1239)
 * So many requests are sent from UI to server after creating match profiles (UIDATIMP-1238)
+* Add column for "Started running" in Data Import Landing page log area (UIDATIMP-1244)
+* Add column for "Started running" in Data Import "View all" page (UIDATIMP-1245)
 
 ### Bugs fixed:
 * Invoice Field Mapping profile: Acq unit and Batch group dropdown lists should be in alphabetical order (UIDATIMP-1236)