+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change history for ui-data-import
 
+## **5.2.4** (in progress)
+
+### Bugs fixed:
+* So many requests are sent from UI to server after creating match profiles (UIDATIMP-1238)
+
 ## [5.2.3](https://github.com/folio-org/ui-data-import/tree/v5.2.3) (2022-08-11)
 
 ### Bugs fixed: